+++ tests/unit/pages/UsersPage.test.js
@@ -6,15 +6,15 @@ import {
   it,
   beforeEach,
 } from 'vitest';
-import IndexPage from 'src/pages/IndexPage.vue';
+import UsersPage from 'pages/UsersPage.vue';
 
 installQuasarPlugin();
 
-describe('Test component: IndexPage', () => {
+describe('Test component: UsersPage', () => {
   let wrapper;
 
   beforeEach(() => {
-    wrapper = shallowMount(IndexPage);
+    wrapper = shallowMount(UsersPage);
   });
 
   it('should mount the component', () => {
